Updates / Fixes on AmChart core libraries.

+++ src/main/java/gwt/material/design/amcharts/client/ui/chart/js/options/AmChartCursor.java
@@ -78,7 +78,7 @@ public class AmChartCursor {
     public String color;
 
     @JsProperty
-    public int cursorAlpha;
+    public double cursorAlpha;
 
     @JsProperty
     public String cursorColor;
